app/requirements.txt
openai
nltk
python-dotenv
mysql-connector-python
